Add files via upload
What is the dataset about?  
The dataset is about the top 10 blockbuster movies per year from 1975 to 2018. The dataset includes information about the title of the movie, the genre it belongs to, studio that produced it, its ranking, IMDB rating and worldwide revenue generated per year.

What are the key questions that you focused on for each: Executives and Middle-level managers?

What is the type of movie that executives and middle-level managers should focus on producing? What is the highest revenue generating genre and rating? How does each studio performed in terms of revenue?